# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: مِرفق API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /properties/{propertyId}/units:
    get:
      operationId: getUnits
      tags: [units]
      summary: List units in a property
      parameters:
        - name: propertyId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of units
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Unit"
    post:
      operationId: createUnit
      tags: [units]
      summary: Create a unit
      parameters:
        - name: propertyId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnitInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
  /units/{id}:
    patch:
      operationId: updateUnit
      tags: [units]
      summary: Update a unit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnitUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
    delete:
      operationId: deleteUnit
      tags: [units]
      summary: Delete a unit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /requests:
    get:
      operationId: getRequests
      tags: [requests]
      summary: List maintenance requests (role-filtered)
      responses:
        "200":
          description: List of requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MaintenanceRequest"
    post:
      operationId: createRequest
      tags: [requests]
      summary: Create a maintenance request (resident only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceRequest"
  /requests/{id}:
    get:
      operationId: getRequest
      tags: [requests]
      summary: Get a maintenance request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceRequest"
        "404":
          description: Not found
    patch:
      operationId: updateRequest
      tags: [requests]
      summary: Update a maintenance request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceRequest"
  /requests/{requestId}/comments:
    get:
      operationId: getRequestComments
      tags: [requests]
      summary: Get comments for a request
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createComment
      tags: [requests]
      summary: Add a comment to a request
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    Unit:
      type: object
      required: [id, number, floor, propertyId]
      properties:
        id:
          type: string
        number:
          type: string
        floor:
          type: integer
        propertyId:
          type: string
        residentId:
          type: ["string", "null"]
        residentName:
          type: ["string", "null"]
    UnitInput:
      type: object
      required: [number, floor]
      properties:
        number:
          type: string
          minLength: 1
        floor:
          type: integer
        residentId:
          type: ["string", "null"]
    UnitUpdate:
      type: object
      properties:
        number:
          type: string
        floor:
          type: integer
        residentId:
          type: ["string", "null"]
    MaintenanceRequest:
      type: object
      required: [id, title, description, category, priority, status, propertyId, unitId, residentId, createdAt, updatedAt]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        category:
          type: string
        priority:
          type: string
        status:
          type: string
        propertyId:
          type: string
        propertyName:
          type: ["string", "null"]
        unitId:
          type: string
        unitNumber:
          type: ["string", "null"]
        residentId:
          type: string
        residentName:
          type: ["string", "null"]
        technicianId:
          type: ["string", "null"]
        technicianName:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    RequestInput:
      type: object
      required: [title, description, category, priority, unitId]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        category:
          type: string
        priority:
          type: string
        unitId:
          type: string
    RequestUpdate:
      type: object
      properties:
        status:
          type: string
        priority:
          type: string
        technicianId:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: string
    Comment:
      type: object
      required: [id, content, requestId, authorId, authorName, createdAt]
      properties:
        id:
          type: string
        content:
          type: string
        requestId:
          type: string
        authorId:
          type: string
        authorName:
          type: string
        createdAt:
          type: string
    CommentInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
...